chore(2024/02): add diff with exploration of more robust part 2
As noted in my commit message for the completion of part 2, I don't
think my solution is entirely correct. This commit adds a diff of some
exploratory work I did after completing today's AoC in an attempt to
improve on this. The work is added as a diff file as it ended up not
fully working in the end but may yet prove useful if the contraption is
required in a future AoC.

The base idea was to use bitmasks for each of the requirements. I.e. one
for the bounds and one for whether a given number in the report was
increasing or decreasing. These would then be possible to combine via
bitwise AND to indicate "good" and "bad" levels. Unfortunately, I failed
 to account for the "bad" level's impact on the direction mask. Example:
 `1 3 1 -9 2` gives a direction mask of `1100?` where `?` is not very
 well-defined.