chore(2024/04): backup solution
Today was kind of rough, and my day has been kind of rough in general...

While my solution eventually worked, it needs cleanup before merge.

At first, I thought I could use this puzzle to implement the modular
solver I was thinking about in yesterday's puzzle (`detectors`
subdirectory). However, this implementation seems to miss some edge case
I have not thought about, and I have so far been unable to find it via
manually counting the occurrences of XMAS.

After a break, I came up with the idea of a char "space" and
considering strings as the characters found by starting at a given
origin and moving in a given direction for a given length. This should
scale well to 3d and higher (which I thought might be part 2.

The implementation ended up a little more verbose than I intended, but
it did work (once I got my x's and y's sorted).

Part 2 turned out to be more about matching 2d vectors than going 3d so
my part 1 solution was not as reusable for this problem. Instead, I
started thinking about digital image processing and the process of
convolving matrices. I think a cleanup of this task should continue
along those lines of thinking as that could easily be reused for part 1.

The actual implementation for part 2 is again a bit verbose and I more
or less forced it to work in my tired state rather than focus on writing
good code.