feat(2024/05): complete part 1
It's interesting switching between kotlin and rust. I find Kotlin to be
more productive as it better matches with my own thinking and intuition.
However, rust tends to be a bit more fun with its focus on performance
and tricky borrowing system. Also, since my work on PrimEvil, connecting
rust code over the JNI, it has become clear to me just how much overhead
there is in the JVM.

My solution for part 1 is rather straight-forward and written in the
main function. I would have liked to optimize the reading a bit so that
I didn't read the file twice. I think my set-based solution is a fun
throwback to the university courses on discrete mathematics. However,
lists do tend to perform better, at least for small datasets, in
practice. I'm sure I would have solved this in a more efficient manner
in rust, just because my mindset would be more on performance and
because the language encourages a more explicit approach.