refactor(2024/06): implement brute force solver for part 2
Why think when stronk? Seriously though, there is definitely a less
resource intensive and more efficient way to solve this. However, for
now this will have to do.

I'm honestly quite happy it was possible to implement the solver as part
of my simulator rather than a complete rewrite.

The original simulator was written while on a bus back to Linköping.
After having spent quite some time writing it up in an awkward position,
I was not thrilled at the prospect of spending additional time writing
part 2 then and there. I also did not have a good idea of how to solve
the problem.

Coming back to the problem later in the day, I was thinking that maybe
I could just throw CPU at the problem as a short-term solution. My idea
was essentially to let the simulation run for X ticks and if it hadn't
stopped by then, assume the guard was looping. Of course, the halting
problem (https://en.wikipedia.org/wiki/Halting_problem) is theoretically
unsolvable, but maybe enough CPU would do the trick in this specific
practice? Unfortunately, AoC was too smart and not even with an X of
`width * height * 1_000` (16.9B) was I able to reach the correct answer.

That's when I realized a much better definition of "looping" is seeing
the guard walk past the same position in the same direction. Not only is
it a better definition, it was quite simple to implement as an extension
to the simulator and makes much better use of CPU resources. My one
remaining discomfort with the solution is the reliance on a separate
"trail" Set. Perhaps I should have gone with a
`Map<Position, Set<Direction>>` instead for more optimal memory usage
and a SSOT.

Of course, the solution still uses brute force to generate each of the
possible map permutations and needs to perform every single simulation
until its end to reach an answer - I'm sure there's an even better
solution out there that does not rely on a simulator but uses more math.