chore(2024/07): complete part 2
Given the size of the answers, I'm very happy that I started this
challenge in python (not having to deal with bigints).

It's quite obvious a major part of the challenge was to handle explosive
permutation counts in an efficient way. Again, python is pretty good at
this since it emphasizes the use of iterators over lists. Still,
performance-wise, I would have preferred using rust or even kotlin.
Calculations without debug logging takes ~30s for part 2. This is a bit
long IMO and could be mitigated using multiprocessing.