feat(2024/09): solve part 1
This is a rather object-oriented approach, which is good as we are
trying to model a "real" thing. This in turn means the defragmentation
and checksum functions can be implemented cleanly and separate from each
other.

A drawback with the object-oriented approach is that we are using more
memory than needed. In fact, it should be possible to work directly on
the input string with just a few helper integers rather than a full
additional list. Such a solution would be able to yield the correct
sequence of file IDs "stored" on the disk, allowing the checksum to be
calculated with minimal memory overhead.